Fixed bbox typecast.

+++ source/LibMultiSense/include/MultiSense/MultiSenseTypes.hh
@@ -2751,7 +2751,7 @@ public:
     
     uint8_t classArray[100];
     float confidenceArray[100];
-    float bboxArray[400];
+    uint16_t bboxArray[400];
     // uint8_t maskArray[57600000];
 };
